refactor: move diffIncoming + DiffResult to store/diff; delete services/utils (B-026)
Final slice of the demote. `diffIncoming` and the `DiffResult` shape
are about diffing two snippet maps — that's a `store/` concern, not
a service helper. New home is `src/store/diff.ts`.

With this slice landed, `services/utils.ts` is empty of useful
exports and gets deleted. Same for its test file (`utils.test.ts`),
whose remaining content was the `diffIncoming` suite — renamed to
`src/store/diff.test.ts`.

Two import sites updated: `ui/components/Modals.ts` (DiffResult),
`ui/components/community/EspansoSection.ts` (diffIncoming).

The `core/install-plan.ts` extraction in PR 2 of this release will
build `buildPackageDiff` on top of `diffIncoming` + `joinKey` rather
than duplicate the logic.

No behaviour changes. tsc clean, 325/325 tests green, bundle
unchanged at 180.6kb.
