+++ src/store/diff.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
-import { diffIncoming } from "./utils";
+import { diffIncoming } from "./diff";
 
-describe("utils.diffIncoming", () => {
+describe("diff.diffIncoming", () => {
     it("splits incoming into added and conflicts", () => {
         const incoming = { a: "1", b: "2", c: "3" };
         const current = { b: "B", d: "4" };
@@ -11,14 +11,14 @@ describe("utils.diffIncoming", () => {
             expect.arrayContaining([
                 { key: "a", value: "1" },
                 { key: "c", value: "3" },
-            ])
+            ]),
         );
         expect(r.added).toHaveLength(2);
 
         expect(r.conflicts).toEqual(
             expect.arrayContaining([
                 { key: "b", incoming: "2", current: "B" },
-            ])
+            ]),
         );
         expect(r.conflicts).toHaveLength(1);
     });
@@ -32,9 +32,3 @@ describe("utils.diffIncoming", () => {
         expect(r.conflicts).toHaveLength(0);
     });
 });
-
-// Trigger-helper tests (normalizeTrigger/isBadTrigger) moved to
-// `src/engine/triggers.test.ts` in 1.1.6 (B-026).
-
-// Key-helper tests (splitKey/joinKey/slugifyGroup/displayGroupTitle)
-// moved to `src/store/keys.test.ts` in 1.1.6 (B-026).
